Switching tests to specs
In an effort to be consistent across projects, I'm opting for
RSpec. We've invested time in RSpec and the speed differences
are negligible for such a small project.
