+++ spec/lib/locabulary/json_creator_spec.rb
@@ -1,18 +1,21 @@
-require_relative '../script/json_creator'
-require 'minitest/autorun'
+require 'spec_helper'
+require 'locabulary/json_creator'
 
-class JsonCreatorTest < MiniTest::Test
-  def test_it_converts_a_hash_to_json_file
-    array_of_data = [
+RSpec.describe Locabulary::JsonCreator do
+  let(:data_fetcher) { ->(document_key) { array_of_data } }
+  let(:array_of_data) do
+    [
       ['Column header 1', 'Column header 2', 'Column header 3', 'Column header 4','Column header 5','Column header 6','Column header 7','Column header 8','Column header 9','Column header 10'],
       ['University of Notre Dame', 'School of Architecture', '', 'College', 'http://architecture.nd.edu/','', '', '','', ''	],
       ['University of Notre Dame', 'Centers and Institutes', 'Center for Building Communities', 'CenterOrInstitute', 'http://buildingcommunities.nd.edu/','','', '', 'University of Notre Dame::School of Architecture']
     ]
+  end
+  subject { described_class.new('dummy', 'administrative_units', data_fetcher) }
 
-    data_fetcher = ->(document_key) {
-      array_of_data
-    }
-
+  it 'will have an output filename' do
+    expect(File.exist?(subject.output_filepath)).to eq(true)
+  end
+  it 'will convert a hash to a json file' do
     expected_output = {
         "predicate_name" => "administrative_units",
         "values" => [
@@ -24,7 +27,7 @@ class JsonCreatorTest < MiniTest::Test
             "description" => "",
             "grouping" => "",
             "affiliation" => "",
-            "presentation_sequence"=> 200,
+            "default_presentation_sequence"=> nil,
             "activated_on" => "2015-07-22",
             "deactivated_on" => nil
           },
@@ -36,15 +39,14 @@ class JsonCreatorTest < MiniTest::Test
             "description" => "",
             "grouping" => "",
             "affiliation" => "University of Notre Dame::School of Architecture",
-            "presentation_sequence" => nil,
+            "default_presentation_sequence" => nil,
             "activated_on" => "2015-07-22",
             "deactivated_on" => nil
           }
         ]
     }
 
-    jcreator = JsonCreator.new('dummy', 'administrative_units', data_fetcher)
-    jcreator.create_or_update
-    assert_equal(jcreator.json_data, JSON.pretty_generate(expected_output))
+    subject.create_or_update
+    expect(subject.json_data).to eq(JSON.pretty_generate(expected_output))
   end
 end
